fixed some zoom issues with candlesticks and renamed/organized the files for candlesticks
