+++ packages/augur-ui/src/modules/market-charts/components/candlestick/candlestick.styles.less
