+++ packages/augur-ui/src/modules/market-charts/components/candlestick/outcome-candlestick.styles.less
@@ -1,6 +1,6 @@
 @import (reference) "~assets/styles/shared";
 
-.MarketOutcomeCandlestick {
+.OutcomeCandlestick {
   background: @color-module-background;
   display: flex;
   flex: 1 100%;
