+++ packages/augur-ui/src/modules/market-charts/containers/candlestick-highchart.ts
@@ -1,12 +1,12 @@
 import { connect } from "react-redux";
 
 import { selectCurrentTimestampInSeconds } from "store/select-state";
-import MarketOutcomeChartsCandlestickHighchart from "modules/market-charts/components/market-outcome-charts--candlestick/market-outcome-charts-candlestick-highchart";
+import CandlestickHighchart from "modules/market-charts/components/candlestick/candlestick-highchart";
 
 const mapStateToProps = (state, ownProps) => ({
   currentTimeInSeconds: selectCurrentTimestampInSeconds(state)
 });
 
 export default connect(mapStateToProps)(
-  MarketOutcomeChartsCandlestickHighchart,
+  CandlestickHighchart,
 );
